# docs/backend-service/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for jira issues and projects analytics.",
        "title": "REST Backend Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/issues": {
            "get": {
                "description": "get list by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "issue"
                ],
                "summary": "Get issues list by project id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.issuesDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    }
                }
            }
        },
        "/issues/fetch": {
            "patch": {
                "description": "fetch issues",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "issue"
                ],
                "summary": "Fetch project's issues locally",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    }
                }
            }
        },
        "/issues/statistics": {
            "get": {
                "description": "get statistics by project ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "issue"
                ],
                "summary": "Get issues statistics on project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.statisticsDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    }
                }
            }
        },
        "/projects": {
            "get": {
                "description": "support pagination for ptojects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Get projects by pages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max number of projects",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "opject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    }
                }
            }
        },
        "/projects/fetch": {
            "patch": {
                "description": "support pagination for ptojects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Get short project info by pages",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max number of projects",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProjectsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "description": "get one fetched projects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Get project info by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "opject"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    }
                }
            }
        },
        "/projects/{id}/fetch": {
            "patch": {
                "description": "fetch project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Fetch project locally",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "408": {
                        "description": "Request Timeout",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    },
                    "504": {
                        "description": "Gateway Timeout",
                        "schema": {
                            "$ref": "#/definitions/errorlib.JSONError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Issue": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "closed_time": {
                    "type": "string"
                },
                "created_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "projectId": {
                    "type": "integer"
                },
                "reporter": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "time_spent": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_time": {
                    "type": "string"
                }
            }
        },
        "dto.PageInfo": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "pageCount": {
                    "type": "integer"
                },
                "projectsCount": {
                    "type": "integer"
                }
            }
        },
        "dto.Project": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ProjectsResponse": {
            "type": "object",
            "properties": {
                "pageInfo": {
                    "$ref": "#/definitions/dto.PageInfo"
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Project"
                    }
                }
            }
        },
        "errorlib.JSONError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer"
                        },
                        "msg": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "v1.issuesDTO": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Issue"
                    }
                },
                "project_id": {
                    "type": "integer"
                }
            }
        },
        "v1.statisticsDTO": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "project_id": {
                    "type": "integer"
                },
                "statistics": {
                    "type": "object",
                    "properties": {
                        "average_time": {
                            "type": "integer"
                        },
                        "closed_number": {
                            "type": "integer"
                        },
                        "opened_number": {
                            "type": "integer"
                        },
                        "total_number": {
                            "type": "integer"
                        },
                        "weekly_average_opened": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}